requirements.txt
sqlparse
google-genai
python-dotenv
langchain[google-genai]
langgraph
faker
